projects/requirements.txt
sanic>=23.12.0
python-dotenv>=1.0.0